fixing response on controller for structure and patient id

+++ mls-donnees-patient/src/main/java/com/phildev/mls/donnees/patient/controller/CoordonneesPatientController.java
@@ -2,7 +2,7 @@ package com.phildev.mls.donnees.patient.controller;
 
 import com.phildev.mls.donnees.patient.exception.PatientException;
 import com.phildev.mls.donnees.patient.exception.PatientNonTrouveException;
-import com.phildev.mls.donnees.patient.exception.StructureNotFoundException;
+import com.phildev.mls.donnees.patient.exception.StructureNonTrouveeException;
 import com.phildev.mls.donnees.patient.model.CoordonneesPatient;
 import com.phildev.mls.donnees.patient.service.CoordonneesPatientService;
 import jakarta.validation.Valid;
@@ -33,9 +33,13 @@ public class CoordonneesPatientController {
      * @return une liste de {@link CoordonneesPatient}
      */
     @GetMapping("/coordonneesPatient/structure/{id}")
-    public List<CoordonneesPatient> getAllCoordonneesPatientByStructureId(@PathVariable("id")@NotNull Integer id){
-        return Optional.ofNullable(coordonneesPatientService.getAllCoordonneesPatientByStructureId(id))
-                .orElseThrow(()-> new StructureNotFoundException("Pas de structure trouvée avec l'id "+id));
+    public List<CoordonneesPatient> getAllCoordonneesPatientByStructureId(@PathVariable("id")@NotNull(message = "l'id ne doit pas être vide") Integer id){
+        List<CoordonneesPatient> coordonneesPatients = coordonneesPatientService.getAllCoordonneesPatientByStructureId(id);
+        if(coordonneesPatients.isEmpty()){
+             LOGGER.error("Pas de structure trouvée avec l'id {}", id);
+             throw new StructureNonTrouveeException("Pas de structure trouvée avec l'id "+id);
+         }
+        return coordonneesPatients;
     }
 
 
@@ -47,9 +51,13 @@ public class CoordonneesPatientController {
      * Si le patient n'est pas trouvé une exception est retourné de type {@link PatientNonTrouveException}
      */
     @GetMapping("/coordonneesPatient/{id}")
-    public Optional<CoordonneesPatient> getCoordonneesPatient(@PathVariable("id") @NotBlank(message = "ne doit pas être vide ou nul") String id ){
-        return Optional.ofNullable(coordonneesPatientService.getAllCoordonneesPatientById(id))
-                .orElseThrow(()-> new PatientNonTrouveException(String.format("Le patient avec l'id %s n'a pas été trouvé", id)));
+    public Optional<CoordonneesPatient> getCoordonneesPatient(@PathVariable("id") @NotBlank(message = "l 'id ne doit pas être vide ou nul") String id ){
+        Optional<CoordonneesPatient> coordonneesPatient =  coordonneesPatientService.getAllCoordonneesPatientById(id);
+        if(coordonneesPatient.isPresent()) {
+            return coordonneesPatientService.getAllCoordonneesPatientById(id);
+        }
+        LOGGER.error("Le patient avec l'id {}} n'a pas été trouvé",id);
+        throw new PatientNonTrouveException(String.format("Le patient avec l'id %s n'a pas été trouvé", id));
     }
 
     /**
@@ -60,15 +68,16 @@ public class CoordonneesPatientController {
      * On renvoie une exception si le patient n'est pas trouvé  de type {@link PatientNonTrouveException}
      */
     @GetMapping("/coordonneesPatient")
-    public CoordonneesPatient getCoordonneesPatient(@RequestParam("nom") @NotBlank(message = "ne doit pas être nul ou vide") String nom,
-                                                    @RequestParam("prenom")@NotBlank(message = "ne doit pas être nul ou vide") String prenom){
-
-        CoordonneesPatient coordonneesPatient = coordonneesPatientService.getAllCoordonneesPatientByNomEtPrenom(nom.toUpperCase(), prenom.toUpperCase());
+    public CoordonneesPatient getCoordonneesPatient(@RequestParam("nom")  @NotBlank(message = "le nom ne doit pas être nul ou vide") String nom,
+                                                    @RequestParam("prenom") @NotBlank(message = "le prénom ne doit pas être nul ou vide") String prenom){
+        String nomEnMajuscules = nom.toUpperCase();
+        String prenomEnMajuscules = prenom.toUpperCase();
+        CoordonneesPatient coordonneesPatient = coordonneesPatientService.getAllCoordonneesPatientByNomEtPrenom(nomEnMajuscules, prenomEnMajuscules);
         if(coordonneesPatient == null){
             LOGGER.error("Aucun patient trouvé avec le nom {} et prénom {}", nom, prenom);
             throw new PatientNonTrouveException(String.format("Aucun patient trouvé avec le nom %s et prénom %s", nom, prenom));
         }
-        return coordonneesPatient;
+        return  coordonneesPatient;
     }
 
     /**
@@ -82,17 +91,18 @@ public class CoordonneesPatientController {
     public ResponseEntity<CoordonneesPatient> createPatient(@Valid @RequestBody CoordonneesPatient coordonneesPatient){
         coordonneesPatient.setNom(coordonneesPatient.getNom().toUpperCase());
         coordonneesPatient.setPrenom(coordonneesPatient.getPrenom().toUpperCase());
-        CoordonneesPatient patientCree = coordonneesPatientService.save(coordonneesPatient);
-        if(patientCree == null){
+        try{
+            CoordonneesPatient patientCree = coordonneesPatientService.save(coordonneesPatient);
+            URI currentUri = ServletUriComponentsBuilder
+                    .fromCurrentRequest()
+                    .buildAndExpand()
+                    .toUri();
+            return ResponseEntity.created(currentUri).body(patientCree);
+        }catch(Exception exception){
+            LOGGER.error("Une erreur est survenue, le patient n'a pas été créé");
             throw new PatientException("Une erreur est survenue, le patient n'a pas été créé");
         }
-        URI currentUri = ServletUriComponentsBuilder
-                .fromCurrentRequest()
-                .buildAndExpand()
-                .toUri();
-        return ResponseEntity.created(currentUri).body(patientCree);
     }
-
 
     /**
      * Cette méthode supprime les coordonnnées d'un patient par son id
@@ -101,10 +111,14 @@ public class CoordonneesPatientController {
      */
     @DeleteMapping(value = "/coordonneesPatient/delete/{id}", produces = "application/json")
     public ResponseEntity<String> deleteCoordonneesPatient(@PathVariable("id") String id){
-       Optional<CoordonneesPatient> coordonneesPatient = Optional.ofNullable(coordonneesPatientService.findById(id)).orElseThrow(() -> new PatientNonTrouveException("Pas de patient trouvé avec cet id :" + id));
-        coordonneesPatient.ifPresent(patient -> LOGGER.info("Suppression du patient {} {}", patient.getPrenom(), patient.getPrenom()));
 
-       coordonneesPatientService.deleteCoordonneesPatient(id);
-       return ResponseEntity.ok().body(String.format("Le patient avec l id %s a bien été supprimé", id));
+       Optional<CoordonneesPatient> coordonneesPatient = coordonneesPatientService.findById(id);
+       if(coordonneesPatient.isPresent()){
+           LOGGER.info("Suppression du patient {} {}", coordonneesPatient.get().getPrenom(), coordonneesPatient.get().getPrenom());
+           coordonneesPatientService.deleteCoordonneesPatient(id);
+           return ResponseEntity.ok().body(String.format("Le patient avec l id %s a bien été supprimé", id));
+        };
+        LOGGER.error("Pas de patient trouvé avec cet id : {}",id);
+        throw new PatientNonTrouveException("Pas de patient trouvé avec cet id : " + id);
     }
 }
